project/requirements.txt
Django==2.2
django-cors-headers==3.2.1
djangorestframework==3.9.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.1
